refactor: change test package structure
