+++ src/test/java/hoomgroom/product/Promo/repository/FixedAmountPromoRepositoryTest.java
@@ -1,8 +1,7 @@
-package hoomgroom.product.Promo.PercentagePromo.repository;
+package hoomgroom.product.Promo.repository;
 
-import hoomgroom.product.Promo.model.Factory.PercentagePromoFactory;
-import hoomgroom.product.Promo.model.PercentagePromo;
-import hoomgroom.product.Promo.repository.PercentagePromoRepository;
+import hoomgroom.product.Promo.model.Factory.FixedAmountPromoFactory;
+import hoomgroom.product.Promo.model.FixedAmountPromo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,73 +19,73 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class PercentagePromoRepositoryTest {
+public class FixedAmountPromoRepositoryTest {
     @Autowired
-    PercentagePromoRepository promoRepository;
+    FixedAmountPromoRepository promoRepository;
 
-    PercentagePromoFactory promoFactory = new PercentagePromoFactory();
+    FixedAmountPromoFactory promoFactory = new FixedAmountPromoFactory();
 
-    List<PercentagePromo> promos;
+    List<FixedAmountPromo> promos;
 
     @BeforeEach
     void setUp() {
         promos = new ArrayList<>();
 
-        PercentagePromo promo1 = promoFactory.createPromo();
-        promo1.setName("BELANJAHEMAT20");
+        FixedAmountPromo promo1 = promoFactory.createPromo();
+        promo1.setName("BELANJAHEMAT20000");
         promo1.setMinimumPurchase(20000000000L);
         promo1.setDescription("Diskon Belanja 20000 dengan minimal pembelian 20 milyar");
-        promo1.setPercentage(20.0);
+        promo1.setDiscountAmount(20000L);
         promos.add(promo1);
 
-        PercentagePromo promo2 = promoFactory.createPromo();
-        promo2.setName("BELANJAHEMAT10");
+        FixedAmountPromo promo2 = promoFactory.createPromo();
+        promo2.setName("BELANJAHEMAT10000");
         promo2.setMinimumPurchase(0L);
-        promo2.setPercentage(10.0);
+        promo2.setDiscountAmount(10000L);
         promos.add(promo2);
 
-        PercentagePromo promo3 = promoFactory.createPromo();
-        promo3.setName("BELANJAHEMAT5");
+        FixedAmountPromo promo3 = promoFactory.createPromo();
+        promo3.setName("BELANJAHEMAT5000");
         promo3.setMinimumPurchase(1000L);
-        promo3.setPercentage(5.0);
+        promo3.setDiscountAmount(5000L);
         promo3.setExpirationDate(LocalDateTime.of(2024, 6, 5, 0, 0));
         promos.add(promo3);
     }
 
     @Test
     public void testSaveCreate() {
-        PercentagePromo promo = promos.get(2);
+        FixedAmountPromo promo = promos.get(2);
         promoRepository.save(promo);
-        Optional<PercentagePromo> findResult = promoRepository.findById(promo.getId());
+        Optional<FixedAmountPromo> findResult = promoRepository.findById(promo.getId());
 
         assertTrue(findResult.isPresent());
         assertEquals(promo.getId(), findResult.get().getId());
         assertEquals(promo.getName(), findResult.get().getName());
         assertEquals(promo.getDescription(), findResult.get().getDescription());
-        assertEquals(promo.getPercentage(), findResult.get().getPercentage());
+        assertEquals(promo.getDiscountAmount(), findResult.get().getDiscountAmount());
         assertEquals(promo.getMinimumPurchase(), findResult.get().getMinimumPurchase());
         assertEquals(promo.getExpirationDate(), findResult.get().getExpirationDate());
     }
 
     @Test
     public void testSaveUpdate() {
-        PercentagePromo promo = promos.get(0);
+        FixedAmountPromo promo = promos.get(0);
         promoRepository.save(promo);
 
-        PercentagePromo newPromo = promoFactory.createPromo();
+        FixedAmountPromo newPromo = promoFactory.createPromo();
         newPromo.setId(promo.getId());
         newPromo.setName(promo.getName());
         newPromo.setDescription(promo.getDescription());
-        newPromo.setPercentage(promo.getPercentage());
+        newPromo.setDiscountAmount(promo.getDiscountAmount());
         newPromo.setMinimumPurchase(promo.getMinimumPurchase());
         newPromo.setExpirationDate(LocalDateTime.of(2024, 6, 5, 0, 0));
 
-        Optional<PercentagePromo> findResult = promoRepository.findById(promo.getId());
+        Optional<FixedAmountPromo> findResult = promoRepository.findById(promo.getId());
         assertTrue(findResult.isPresent());
         assertEquals(promo.getId(), findResult.get().getId());
         assertEquals(promo.getName(), findResult.get().getName());
         assertEquals(promo.getDescription(), findResult.get().getDescription());
-        assertEquals(promo.getPercentage(), findResult.get().getPercentage());
+        assertEquals(promo.getDiscountAmount(), findResult.get().getDiscountAmount());
         assertEquals(promo.getMinimumPurchase(), findResult.get().getMinimumPurchase());
         assertEquals(promo.getExpirationDate(), findResult.get().getExpirationDate());
     }
@@ -95,14 +94,14 @@ public class PercentagePromoRepositoryTest {
     public void testFindByIdIfIdFound() {
         promoRepository.saveAll(promos);
 
-        PercentagePromo target = promos.get(0);
+        FixedAmountPromo target = promos.get(0);
 
-        Optional<PercentagePromo> findResult = promoRepository.findById(target.getId());
+        Optional<FixedAmountPromo> findResult = promoRepository.findById(target.getId());
         assertTrue(findResult.isPresent());
         assertEquals(target.getId(), findResult.get().getId());
         assertEquals(target.getName(), findResult.get().getName());
         assertEquals(target.getDescription(), findResult.get().getDescription());
-        assertEquals(target.getPercentage(), findResult.get().getPercentage());
+        assertEquals(target.getDiscountAmount(), findResult.get().getDiscountAmount());
         assertEquals(target.getMinimumPurchase(), findResult.get().getMinimumPurchase());
     }
 
@@ -110,7 +109,7 @@ public class PercentagePromoRepositoryTest {
     public void testFindByIdIfIdNotFound() {
         promoRepository.saveAll(promos);
 
-        Optional<PercentagePromo> findResult = promoRepository.findById(UUID.fromString("0f14d0ab-9605-4a62-a9e4-5ed26688389b"));
+        Optional<FixedAmountPromo> findResult = promoRepository.findById(UUID.fromString("0f14d0ab-9605-4a62-a9e4-5ed26688389b"));
         assertTrue(findResult.isEmpty());
     }
 
@@ -118,14 +117,14 @@ public class PercentagePromoRepositoryTest {
     public void testFindByNameFound() {
         promoRepository.saveAll(promos);
 
-        PercentagePromo target = promos.get(0);
+        FixedAmountPromo target = promos.get(0);
 
-        List<PercentagePromo> findResult = promoRepository.findByName(target.getName());
+        List<FixedAmountPromo> findResult = promoRepository.findByName(target.getName());
         assertTrue(findResult.contains(target));
         assertEquals(target.getId(), findResult.getFirst().getId());
         assertEquals(target.getName(), findResult.getFirst().getName());
         assertEquals(target.getDescription(), findResult.getFirst().getDescription());
-        assertEquals(target.getPercentage(), findResult.getFirst().getPercentage());
+        assertEquals(target.getDiscountAmount(), findResult.getFirst().getDiscountAmount());
         assertEquals(target.getMinimumPurchase(), findResult.getFirst().getMinimumPurchase());
         assertEquals(target.getExpirationDate(), findResult.getFirst().getExpirationDate());
     }
@@ -134,7 +133,7 @@ public class PercentagePromoRepositoryTest {
     public void testFindByNameNotFound() {
         promoRepository.saveAll(promos);
 
-        List<PercentagePromo> findResult = promoRepository.findByName("name_not_found");
+        List<FixedAmountPromo> findResult = promoRepository.findByName("name_not_found");
         assertTrue(findResult.isEmpty());
     }
 
@@ -142,15 +141,15 @@ public class PercentagePromoRepositoryTest {
     public void testFindByExpirationDateBeforeFound() {
         promoRepository.saveAll(promos);
 
-        PercentagePromo target = promos.get(2);
+        FixedAmountPromo target = promos.get(2);
 
-        List<PercentagePromo> findResult = promoRepository.findByExpirationDateBefore(
+        List<FixedAmountPromo> findResult = promoRepository.findByExpirationDateBefore(
                 LocalDateTime.of(2024, 6, 5, 0, 0));
         assertTrue(findResult.contains(target));
         assertEquals(target.getId(), findResult.getFirst().getId());
         assertEquals(target.getName(), findResult.getFirst().getName());
         assertEquals(target.getDescription(), findResult.getFirst().getDescription());
-        assertEquals(target.getPercentage(), findResult.getFirst().getPercentage());
+        assertEquals(target.getDiscountAmount(), findResult.getFirst().getDiscountAmount());
         assertEquals(target.getMinimumPurchase(), findResult.getFirst().getMinimumPurchase());
         assertEquals(target.getExpirationDate(), findResult.getFirst().getExpirationDate());
     }
@@ -159,33 +158,33 @@ public class PercentagePromoRepositoryTest {
     public void testFindByExpirationDateBeforeNotFound() {
         promoRepository.saveAll(promos);
 
-        List<PercentagePromo> findResult = promoRepository.findByExpirationDateBefore(
+        List<FixedAmountPromo> findResult = promoRepository.findByExpirationDateBefore(
                 LocalDateTime.of(2021, 6, 5, 0, 0));
         assertTrue(findResult.isEmpty());
     }
 
     @Test
-    public void testFindByPercentageFound() {
+    public void testFindByDiscountAmountFound() {
         promoRepository.saveAll(promos);
 
-        PercentagePromo target = promos.get(2);
+        FixedAmountPromo target = promos.get(2);
 
-        List<PercentagePromo> findResult = promoRepository.findByPercentageBetween(0.0, 6.0);
+        List<FixedAmountPromo> findResult = promoRepository.findByDiscountAmountBetween(0L, 9000L);
         assertEquals(1, findResult.size());
         assertEquals(target.getId(), findResult.getFirst().getId());
         assertEquals(target.getName(), findResult.getFirst().getName());
         assertEquals(target.getDescription(), findResult.getFirst().getDescription());
-        assertEquals(target.getPercentage(), findResult.getFirst().getPercentage());
+        assertEquals(target.getDiscountAmount(), findResult.getFirst().getDiscountAmount());
         assertEquals(target.getMinimumPurchase(), findResult.getFirst().getMinimumPurchase());
         assertEquals(target.getExpirationDate(), findResult.getFirst().getExpirationDate());
     }
 
     @Test
-    public void testFindByPercentageNotFound() {
+    public void testFindByDiscountAmountNotFound() {
         promoRepository.saveAll(promos);
 
-        List<PercentagePromo> findResult = promoRepository
-                .findByPercentageBetween(1000000000.0, 3000000000.0);
+        List<FixedAmountPromo> findResult = promoRepository
+                .findByDiscountAmountBetween(1000000000L, 3000000000L);
 
         assertTrue(findResult.isEmpty());
     }
